pyproject.toml
[project]
name = "message-beam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-beam[gcp]>=2.70.0",
    "flask>=3.1.2",
    "google-cloud-pubsub>=2.35.0",
    "requests>=2.32.5",
]
